[chore]  entity 파일 분리
